% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.tidycuminc}
\alias{autoplot.tidycuminc}
\title{Plot cuminc Estimates}
\usage{
\method{autoplot}{tidycuminc}(object, ...)
}
\arguments{
\item{object}{object of class 'cuminc'}

\item{...}{not used}
}
\description{
Function uses the result from \code{tidy(object)} to create figure.
}
\examples{
cuminc(Surv(ttdeath, death_cr) ~ trt, trial) \%>\%
  autoplot()
}
