% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{base_methods_cuminc}
\alias{base_methods_cuminc}
\alias{model.frame.tidycuminc}
\alias{model.matrix.tidycuminc}
\title{Functions for tidycuminc objects}
\usage{
\method{model.frame}{tidycuminc}(formula, ...)

\method{model.matrix}{tidycuminc}(object, ...)
}
\arguments{
\item{formula}{a formula}

\item{...}{not used}

\item{object}{a tidycuminc object}
}
\description{
Functions for tidycuminc objects
}
