% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{broom_methods}
\alias{broom_methods}
\alias{tidy.tidycrr}
\alias{glance.tidycrr}
\alias{augment.tidycrr}
\title{Broom methods for tidycmprsk objects}
\usage{
\method{tidy}{tidycrr}(x, ...)

\method{glance}{tidycrr}(x, ...)

\method{augment}{tidycrr}(x, quantiles = seq(0, 1, 0.25), new_data, ...)
}
\arguments{
\item{x}{placeholder}

\item{...}{not used}

\item{quantiles}{placeholder}

\item{new_data}{placeholder}
}
\value{
a tibble
}
\description{
Broom methods for tidycmprsk objects
}
\concept{tidycrr tidiers}
