% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crSurv.R
\name{crSurv}
\alias{crSurv}
\title{Competing Risks Outcome Definition}
\usage{
crSurv(time, event)
}
\arguments{
\item{time}{vector of failure/censoring times}

\item{event}{vector with a unique code for each failure type and a separate
code for censored observations}
}
\value{

}
\description{
I think this fn should also somehow define the event of interest and the
censoring code somehow....? as an attribute?
}
\examples{
# add example
}
