% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuminc.R
\name{cuminc}
\alias{cuminc}
\alias{cuminc.formula}
\alias{cuminc.default}
\title{Competing Risks Regression}
\usage{
\method{cuminc}{formula}(formula, data, strata, rho = 0, ...)

cuminc(x, ...)

\method{cuminc}{default}(x, ...)
}
\arguments{
\item{formula}{formula with \code{Surv()} on LHS and covariates on RHS.}

\item{data}{data frame}

\item{...}{passed to methods}

\item{x}{input object}
}
\value{
tidycuminc object
}
\description{
Competing Risks Regression
}
\examples{
cuminc(Surv(ttdeath, death_cr) ~ trt, trial)
}
