% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{predict.tidycrr}
\alias{predict.tidycrr}
\title{Predict}
\usage{
\method{predict}{tidycrr}(object, new_data = NULL, quantiles = seq(0, 1, 0.25), ...)
}
\arguments{
\item{object}{placeholder}

\item{new_data}{placeholder}

\item{quantiles}{placeholder}

\item{...}{not used}
}
\description{
Predict
}
