% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.tidycrr}
\alias{print.tidycuminc}
\title{Print crr object}
\usage{
\method{print}{tidycrr}(x, ...)

\method{print}{tidycuminc}(x, ...)
}
\arguments{
\item{x}{tidycrr object}

\item{...}{not used}
}
\description{
Print crr object
}
\keyword{internal}
