% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycmprsk-package.R
\docType{package}
\name{tidycmprsk-package}
\alias{tidycmprsk}
\alias{tidycmprsk-package}
\title{tidycmprsk: Tidy Implementation of Competing Risks Regression}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mskcc-epi-bio.github.io/tidycmprsk}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Teng Fei
}

}
\keyword{internal}
