% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.tidycuminc}
\alias{autoplot.tidycuminc}
\title{Plot Cumulative Incidence Estimates}
\usage{
\method{autoplot}{tidycuminc}(
  object,
  outcomes = NULL,
  conf.int = FALSE,
  conf.level = 0.95,
  aes = NULL,
  ...
)
}
\arguments{
\item{object, outcomes, aes, conf.int, conf.level, ...}{\strong{DEPRECATED}}
}
\description{
\strong{DEPRECATED} Use \code{\link[ggsurvfit:ggcuminc]{ggsurvfit::ggcuminc()}} instead.
}
\keyword{internal}
