% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{base_methods}
\alias{base_methods}
\alias{model.matrix.tidycrr}
\alias{model.frame.tidycrr}
\alias{model.frame.tidycuminc}
\alias{predict.tidycrr}
\alias{terms.tidycrr}
\alias{coef.tidycrr}
\title{Methods for tidycmprsk objects}
\usage{
\method{model.matrix}{tidycrr}(object, ...)

\method{model.frame}{tidycrr}(formula, ...)

\method{model.frame}{tidycuminc}(formula, ...)

\method{predict}{tidycrr}(object, new_data = NULL, quantiles = seq(0, 1, 0.25), ...)

\method{terms}{tidycrr}(x, ...)

\method{coef}{tidycrr}(object, ...)
}
\arguments{
\item{...}{not used}

\item{new_data}{a data frame}

\item{quantiles}{vector of quantiles}

\item{x, object, formula}{a tidycmprsk object}
}
\description{
Methods for tidycmprsk objects

Predict
}
