% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{broom_methods}
\alias{broom_methods}
\alias{tidy.tidycrr}
\alias{glance.tidycrr}
\alias{augment.tidycrr}
\alias{tidy.tidycuminc}
\alias{glance.tidycuminc}
\title{Broom methods for tidycmprsk objects}
\usage{
\method{tidy}{tidycrr}(x, exponentiate = FALSE, conf.int = FALSE, conf.level = 0.95, ...)

\method{glance}{tidycrr}(x, ...)

\method{augment}{tidycrr}(x, times = NULL, probs = NULL, newdata = NULL, ...)

\method{tidy}{tidycuminc}(x, conf.int = FALSE, conf.level = 0.95, times = NULL, ...)

\method{glance}{tidycuminc}(x, ...)
}
\arguments{
\item{x}{placeholder}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
the coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults
to \code{FALSE}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{not used}

\item{times}{vector of numeric time points where risk estimates will be shown.
Default it to use all observed times.}

\item{probs}{probs}

\item{newdata}{A \code{base::data.frame()} or \code{tibble::tibble()} containing all
the original predictors used to create x. Defaults to \code{NULL}.}
}
\value{
a tibble
}
\description{
Broom methods for tidycmprsk objects
}
\concept{tidycrr tidiers}
