% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_methods.R
\name{broom_methods_cuminc}
\alias{broom_methods_cuminc}
\alias{tidy.tidycuminc}
\alias{glance.tidycuminc}
\title{Broom methods for tidy cuminc objects}
\usage{
\method{tidy}{tidycuminc}(x, times = NULL, conf.int = TRUE, conf.level = 0.95, ...)

\method{glance}{tidycuminc}(x, ...)
}
\arguments{
\item{x}{object of class 'tidycuminc'}

\item{times}{Numeric vector of times to obtain risk estimates at}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Broom methods for tidy cuminc objects
}
\section{\code{tidy()} data frame}{


The returned \code{tidy()} data frame returns the following columns:\tabular{ll}{
   \strong{Column Name} \tab \strong{Description} \cr
   \code{outcome} \tab Competing Event Outcome \cr
   \code{time} \tab Numeric follow-up time \cr
   \code{estimate} \tab Risk estimate \cr
   \code{std.error} \tab Standard Error \cr
   \code{n.risk} \tab Number at risk at the specified time \cr
   \code{n.event} \tab Number of events at specified time \cr
   \code{n.censor} \tab Number of censored observations at specified time \cr
   \code{cumulative.event} \tab Cumulative number of events at specified time \cr
   \code{cumulative.censor} \tab Cumulative number of censored observations at specified time \cr
}
}

\section{Confidence intervals}{


The confidence intervals for cumulative incidence estimates use the recommended method in
\emph{Competing Risks: A Practical Perspective} by Melania Pintilie.

\deqn{x^exp(±z * se / (x * log(x))),}

where x is the cumulative incidence estimate and z = 1.96 is the 97.5th percentile of the standard normal distribution.
}

\examples{
cuminc <- cuminc(Surv(ttdeath, death_cr) ~ trt, trial)

tidy(cuminc)

glance(cuminc)
}
\seealso{
Other cuminc() functions: 
\code{\link{autoplot.tidycuminc}()},
\code{\link{cuminc}()}
}
\concept{cuminc tidiers}
\concept{cuminc() functions}
\concept{tidycuminc tidiers}
