% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr.R
\name{crr}
\alias{crr}
\alias{crr.default}
\alias{crr.formula}
\alias{print.tidycrr}
\alias{model.matrix.tidycrr}
\alias{model.frame.tidycrr}
\alias{tidy.tidycrr}
\alias{glance.tidycrr}
\alias{predict.tidycrr}
\alias{augment.tidycrr}
\title{Competing Risks Regression}
\usage{
crr(x, ...)

\method{crr}{default}(x, ...)

\method{crr}{formula}(formula, data, failcode = 1, ...)

\method{print}{tidycrr}(x, ...)

\method{model.matrix}{tidycrr}(object, ...)

\method{model.frame}{tidycrr}(formula, ...)

\method{tidy}{tidycrr}(object, ...)

\method{glance}{tidycrr}(object, ...)

\method{predict}{tidycrr}(object, new_data = NULL, quantiles = seq(0, 1, 0.25), ...)

\method{augment}{tidycrr}(object, quantiles = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{placeholder}

\item{...}{placeholder}

\item{formula}{placeholder}

\item{data}{placeholder}

\item{failcode}{placeholder}

\item{object}{placeholder}

\item{new_data}{placeholder}

\item{quantiles}{placeholder}
}
\value{
tidycrr object
}
\description{
Competing Risks Regression
}
\examples{
# ADD EXAMPLE!
}
\concept{tidycrr tidiers}
