% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr.R
\name{crr}
\alias{crr}
\alias{crr.default}
\alias{crr.formula}
\alias{print.tidycmprsk}
\alias{model.matrix.tidycmprsk}
\title{Competing Risks Regression}
\usage{
crr(x, ...)

\method{crr}{default}(x, ...)

\method{crr}{formula}(formula, data, failcode = 1, ...)

\method{print}{tidycmprsk}(object)

\method{model.matrix}{tidycmprsk}(object, ...)
}
\arguments{
\item{x}{placeholder}

\item{...}{placeholder}
}
\value{

}
\description{
Competing Risks Regression
}
\examples{
# ADD EXAMPLE!
}
