% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr.R
\name{crr}
\alias{crr}
\alias{crr.default}
\alias{crr.matrix}
\alias{crr.formula}
\title{Competing Risks Regression}
\usage{
crr(x, ...)

\method{crr}{default}(x, ...)

\method{crr}{matrix}(x, y, ...)

\method{crr}{formula}(formula, data, ...)
}
\arguments{
\item{x}{placeholder}

\item{...}{placeholder}
}
\value{

}
\description{
Competing Risks Regression
}
\examples{
# ADD EXAMPLE!
}
