% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr.R
\name{crr}
\alias{crr}
\alias{crr.default}
\alias{crr.formula}
\title{Competing Risks Regression}
\usage{
crr(x, ...)

\method{crr}{default}(x, ...)

\method{crr}{formula}(formula, data, failcode = NULL, ...)
}
\arguments{
\item{x}{input object}

\item{...}{passed to methods}

\item{formula}{formula with \code{Surv()} on LHS and covariates on RHS.}

\item{data}{data frame}

\item{failcode}{indicates event of interest. Default is \code{NULL} and event
of interest is inferred from the data.}
}
\value{
tidycrr object
}
\description{
Competing Risks Regression
}
\examples{
crr(Surv(ttdeath, death_cr) ~ age, trial)
}
