% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{autoplot.tidycuminc}
\title{Deprecated}
\usage{
\method{autoplot}{tidycuminc}(...)
}
\arguments{
\item{...}{not used}
}
\value{
Misc.
}
\description{
\strong{DEPRECATED} functions.
}
\examples{
# DO NOT USE DEPRECATED FUNCTIONS
}
\keyword{internal}
