% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{Surv}
\alias{autoplot}
\alias{tidy}
\alias{glance}
\alias{augment}
\alias{add_n}
\alias{add_nevent}
\alias{add_p}
\alias{inline_text}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{augment}}, \code{\link[broom:reexports]{glance}}, \code{\link[broom:reexports]{tidy}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{gtsummary}{\code{\link[gtsummary]{add_n}}, \code{\link[gtsummary]{add_nevent}}, \code{\link[gtsummary]{add_p}}, \code{\link[gtsummary]{inline_text}}}

  \item{survival}{\code{\link[survival]{Surv}}}
}}

