% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycmprsk-package.R
\docType{package}
\name{tidycmprsk-package}
\alias{tidycmprsk}
\alias{tidycmprsk-package}
\title{tidycmprsk: Tidy Implementation of Competing Risks Regression}
\description{
Provides an improved interface for working with the competing risk endpoints. The package wraps the 'cmprsk' package, and exports functions for univariate cumulative incidence estimates and competing risk regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mskcc-epi-bio.github.io/tidycmprsk}
  \item Report bugs at \url{https://mskcc-epi-bio.github.io/tidycmprsk/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Teng Fei (\href{https://orcid.org/0000-0001-7888-1715}{ORCID})
}

}
\keyword{internal}
