% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycmprsk-package.R
\docType{package}
\name{tidycmprsk-package}
\alias{tidycmprsk}
\alias{tidycmprsk-package}
\title{tidycmprsk: Tidy Implementation of Competing Risks Regression}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mskcc-epi-bio.github.io/tidycmprsk}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
