% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycmprsk-package.R
\docType{package}
\name{tidycmprsk-package}
\alias{tidycmprsk}
\alias{tidycmprsk-package}
\title{tidycmprsk: Competing Risks Estimation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an intuitive interface for working with the competing risk endpoints. The package wraps the 'cmprsk' package, and exports functions for univariate cumulative incidence estimates and competing risk regression. Methods follow those introduced in Fine and Gray (1999) \doi{10.1002/sim.7501}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mskcc-epi-bio.github.io/tidycmprsk/}
  \item \url{https://github.com/MSKCC-Epi-Bio/tidycmprsk}
  \item Report bugs at \url{https://github.com/MSKCC-Epi-Bio/tidycmprsk/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Teng Fei (\href{https://orcid.org/0000-0001-7888-1715}{ORCID})
}

}
\keyword{internal}
